+++ sdk/storage/azure-storage-file-datalake/src/main/java/com/azure/storage/file/datalake/implementation/models/PathsUndeleteHeaders.java
@@ -10,16 +10,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 import java.time.OffsetDateTime;
 
-/** The PathsDeleteHeaders model. */
+/** The PathsUndeleteHeaders model. */
 @JacksonXmlRootElement(localName = "null")
 @Fluent
-public final class PathsDeleteHeaders {
-    /*
-     * The x-ms-deletion-id property.
-     */
-    @JsonProperty(value = "x-ms-deletion-id")
-    private String xMsDeletionId;
-
+public final class PathsUndeleteHeaders {
     /*
      * The x-ms-version property.
      */
@@ -33,36 +27,16 @@ public final class PathsDeleteHeaders {
     private String xMsRequestId;
 
     /*
-     * The x-ms-continuation property.
+     * The x-ms-client-request-id property.
      */
-    @JsonProperty(value = "x-ms-continuation")
-    private String xMsContinuation;
+    @JsonProperty(value = "x-ms-client-request-id")
+    private String xMsClientRequestId;
 
     /*
      * The Date property.
      */
     @JsonProperty(value = "Date")
     private DateTimeRfc1123 dateProperty;
-
-    /**
-     * Get the xMsDeletionId property: The x-ms-deletion-id property.
-     *
-     * @return the xMsDeletionId value.
-     */
-    public String getXMsDeletionId() {
-        return this.xMsDeletionId;
-    }
-
-    /**
-     * Set the xMsDeletionId property: The x-ms-deletion-id property.
-     *
-     * @param xMsDeletionId the xMsDeletionId value to set.
-     * @return the PathsDeleteHeaders object itself.
-     */
-    public PathsDeleteHeaders setXMsDeletionId(String xMsDeletionId) {
-        this.xMsDeletionId = xMsDeletionId;
-        return this;
-    }
 
     /**
      * Get the xMsVersion property: The x-ms-version property.
@@ -77,9 +51,9 @@ public final class PathsDeleteHeaders {
      * Set the xMsVersion property: The x-ms-version property.
      *
      * @param xMsVersion the xMsVersion value to set.
-     * @return the PathsDeleteHeaders object itself.
+     * @return the PathsUndeleteHeaders object itself.
      */
-    public PathsDeleteHeaders setXMsVersion(String xMsVersion) {
+    public PathsUndeleteHeaders setXMsVersion(String xMsVersion) {
         this.xMsVersion = xMsVersion;
         return this;
     }
@@ -97,30 +71,30 @@ public final class PathsDeleteHeaders {
      * Set the xMsRequestId property: The x-ms-request-id property.
      *
      * @param xMsRequestId the xMsRequestId value to set.
-     * @return the PathsDeleteHeaders object itself.
+     * @return the PathsUndeleteHeaders object itself.
      */
-    public PathsDeleteHeaders setXMsRequestId(String xMsRequestId) {
+    public PathsUndeleteHeaders setXMsRequestId(String xMsRequestId) {
         this.xMsRequestId = xMsRequestId;
         return this;
     }
 
     /**
-     * Get the xMsContinuation property: The x-ms-continuation property.
+     * Get the xMsClientRequestId property: The x-ms-client-request-id property.
      *
-     * @return the xMsContinuation value.
+     * @return the xMsClientRequestId value.
      */
-    public String getXMsContinuation() {
-        return this.xMsContinuation;
+    public String getXMsClientRequestId() {
+        return this.xMsClientRequestId;
     }
 
     /**
-     * Set the xMsContinuation property: The x-ms-continuation property.
+     * Set the xMsClientRequestId property: The x-ms-client-request-id property.
      *
-     * @param xMsContinuation the xMsContinuation value to set.
-     * @return the PathsDeleteHeaders object itself.
+     * @param xMsClientRequestId the xMsClientRequestId value to set.
+     * @return the PathsUndeleteHeaders object itself.
      */
-    public PathsDeleteHeaders setXMsContinuation(String xMsContinuation) {
-        this.xMsContinuation = xMsContinuation;
+    public PathsUndeleteHeaders setXMsClientRequestId(String xMsClientRequestId) {
+        this.xMsClientRequestId = xMsClientRequestId;
         return this;
     }
 
@@ -140,9 +114,9 @@ public final class PathsDeleteHeaders {
      * Set the dateProperty property: The Date property.
      *
      * @param dateProperty the dateProperty value to set.
-     * @return the PathsDeleteHeaders object itself.
+     * @return the PathsUndeleteHeaders object itself.
      */
-    public PathsDeleteHeaders setDateProperty(OffsetDateTime dateProperty) {
+    public PathsUndeleteHeaders setDateProperty(OffsetDateTime dateProperty) {
         if (dateProperty == null) {
             this.dateProperty = null;
         } else {